requirements.txt
Flask==3.0.0
fuzzywuzzy==0.18.0
Levenshtein==0.23.0
nltk==3.8.1
python-Levenshtein==0.23.0
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
spacy==3.7.2
openai==0.28.1
python-dotenv==0.21.0
